requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.1
cloudinary==1.36.0
dj-database-url==2.1.0
Django==4.2.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.0
django-environ==0.11.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
idna==3.4
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
whitenoise==6.6.0
